fix queue by removing it altogether
